Added drink prices for all the pubs in the Eyre Square region, moved some code around to avoid bulk code and sorted some bugs out within the code.
